Optimisation, renommage, Full Autowired à la place de la méthode constructeur
