+++ src/main/java/com/groupe2/redline/entities/Site.java
@@ -1,6 +1,5 @@
 package com.groupe2.redline.entities;
 
-import com.groupe2.redline.controllers.dto.SiteDTO;
 import jakarta.persistence.*;
 
 import java.util.Set;
