+++ src/main/java/com/groupe2/redline/repositories/SiteRepository.java
@@ -1,7 +1,6 @@
-package com.groupe2.redline.repository;
+package com.groupe2.redline.repositories;
 
 import com.groupe2.redline.entities.Site;
-import org.apache.juli.logging.Log;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
