+++ src/main/java/com/groupe2/redline/services/DemandeService.java
@@ -1,7 +1,7 @@
 package com.groupe2.redline.services;
 
 import com.groupe2.redline.entities.Demande;
-import com.groupe2.redline.repository.DemandeRepository;
+import com.groupe2.redline.repositories.DemandeRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
