+++ src/main/java/com/groupe2/redline/services/SiteService.java
@@ -1,13 +1,14 @@
 package com.groupe2.redline.services;
 
-import com.groupe2.redline.controllers.dto.SiteDTO;
-import com.groupe2.redline.controllers.mappers.SiteMapper;
+import com.groupe2.redline.dto.SiteDto;
+import com.groupe2.redline.mappers.SiteMapper;
 import com.groupe2.redline.entities.Site;
 import com.groupe2.redline.exceptions.SiteDejaActifException;
 import com.groupe2.redline.exceptions.SiteDejaInactifException;
-import com.groupe2.redline.repository.SiteRepository;
+import com.groupe2.redline.repositories.SiteRepository;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
@@ -18,14 +19,11 @@ import java.util.Optional;
 @Transactional
 public class SiteService {
 
-    private final SiteRepository siteRepository;
+    @Autowired
+    private SiteRepository siteRepository;
 
-    private final SiteMapper siteMapper;
-
-    public SiteService(SiteRepository siteRepository, SiteMapper siteMapper) {
-        this.siteRepository = siteRepository;
-        this.siteMapper = siteMapper;
-    }
+    @Autowired
+    private SiteMapper siteMapper;
 
     public List<Site> getAllSites() {
         List<Site> sites = this.siteRepository.findAll(Sort.by(Sort.Direction.ASC, "libelle"));
@@ -42,13 +40,13 @@ public class SiteService {
         return siteRepository.findById(id);
     }
 
-    public Site editSite(Long id, SiteDTO siteDTO) throws EntityNotFoundException {
+    public Site editSite(Long id, SiteDto siteDTO) throws EntityNotFoundException {
         Optional<Site> editingSite = siteRepository.findById(id);
         if (editingSite.isEmpty()) {
             throw new EntityNotFoundException("Le site avec l'ID " + id + " n'existe pas.");
         }
 
-        Site updatedSite = siteMapper.editSitefromDTO(editingSite.get(), siteDTO);
+        Site updatedSite = siteMapper.editSiteFromDTO(editingSite.get(), siteDTO);
         Site savedSite = siteRepository.save(updatedSite);
 
         return savedSite;
